requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.54.0
anyio==4.9.0
async-timeout==4.0.3
asyncer==0.0.7
attrs==25.3.0
backoff==2.2.1
bidict==0.23.1
certifi==2025.6.15
chainlit==2.5.5
charset-normalizer==3.4.2
chevron==0.14.0
click==8.2.1
colorama==0.4.6
dataclasses-json==0.6.7
Deprecated==1.2.18
distro==1.9.0
exceptiongroup==1.3.0
fastapi==0.115.13
filelock==3.18.0
filetype==1.2.0
frozenlist==1.7.0
fsspec==2025.5.1
galileo==1.4.0
galileo-core==3.44.0
googleapis-common-protos==1.70.0
grpcio==1.73.0
h11==0.16.0
hf-xet==1.1.4
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.33.0
idna==3.10
importlib_metadata==8.6.1
inflection==0.5.1
Jinja2==3.1.6
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.25
langchain-core==0.3.65
langchain-openai==0.3.24
langchain-text-splitters==0.3.8
langsmith==0.3.45
Lazify==0.4.0
literalai==0.1.201
MarkupSafe==3.0.2
marshmallow==3.26.1
mcp==1.9.4
monotonic==1.6
multidict==6.5.0
mypy_extensions==1.1.0
nest-asyncio==1.6.0
numpy==2.2.6
openai==1.88.0
opentelemetry-api==1.31.1
opentelemetry-exporter-otlp==1.31.1
opentelemetry-exporter-otlp-proto-common==1.31.1
opentelemetry-exporter-otlp-proto-grpc==1.31.1
opentelemetry-exporter-otlp-proto-http==1.31.1
opentelemetry-instrumentation==0.52b1
opentelemetry-instrumentation-alephalpha==0.40.11
opentelemetry-instrumentation-anthropic==0.40.11
opentelemetry-instrumentation-bedrock==0.40.11
opentelemetry-instrumentation-chromadb==0.40.11
opentelemetry-instrumentation-cohere==0.40.11
opentelemetry-instrumentation-crewai==0.40.11
opentelemetry-instrumentation-google-generativeai==0.40.11
opentelemetry-instrumentation-groq==0.40.11
opentelemetry-instrumentation-haystack==0.40.11
opentelemetry-instrumentation-lancedb==0.40.11
opentelemetry-instrumentation-langchain==0.40.11
opentelemetry-instrumentation-llamaindex==0.40.11
opentelemetry-instrumentation-logging==0.52b1
opentelemetry-instrumentation-marqo==0.40.11
opentelemetry-instrumentation-mcp==0.40.11
opentelemetry-instrumentation-milvus==0.40.11
opentelemetry-instrumentation-mistralai==0.40.11
opentelemetry-instrumentation-ollama==0.40.11
opentelemetry-instrumentation-openai==0.40.11
opentelemetry-instrumentation-pinecone==0.40.11
opentelemetry-instrumentation-qdrant==0.40.11
opentelemetry-instrumentation-replicate==0.40.11
opentelemetry-instrumentation-requests==0.52b1
opentelemetry-instrumentation-sagemaker==0.40.11
opentelemetry-instrumentation-sqlalchemy==0.52b1
opentelemetry-instrumentation-threading==0.52b1
opentelemetry-instrumentation-together==0.40.11
opentelemetry-instrumentation-transformers==0.40.11
opentelemetry-instrumentation-urllib3==0.52b1
opentelemetry-instrumentation-vertexai==0.40.11
opentelemetry-instrumentation-watsonx==0.40.11
opentelemetry-instrumentation-weaviate==0.40.11
opentelemetry-proto==1.31.1
opentelemetry-sdk==1.31.1
opentelemetry-semantic-conventions==0.52b1
opentelemetry-semantic-conventions-ai==0.4.9
opentelemetry-util-http==0.52b1
orjson==3.10.18
packaging==24.2
posthog==3.25.0
propcache==0.3.2
protobuf==5.29.5
pydantic==2.11.7
pydantic-settings==2.9.1
pydantic_core==2.33.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-engineio==4.12.2
python-multipart==0.0.18
python-socketio==5.13.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
sse-starlette==2.3.6
starlette==0.41.3
syncer==2.0.3
tenacity==9.1.2
tiktoken==0.9.0
tokenizers==0.21.1
tomli==2.2.1
tqdm==4.67.1
traceloop-sdk==0.40.11
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
uptrace==1.31.0
urllib3==2.5.0
uvicorn==0.34.3
uvloop==0.21.0
watchfiles==0.20.0
wrapt==1.17.2
wsproto==1.2.0
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0
